Clean up before demo

+++ TimeSeriesWebDemo/Models/DemoTimeSeries.cs
@@ -5,6 +5,6 @@
     {
         public string Id { get; set; }
         public string Source { get; set; }
-        public double[] Values { get; set; }
+        public decimal[] Values { get; set; }
     }
 }
